requirements.txt
Django==3.1.7
djangorestframework==3.12.2
psycopg2-binary==2.8.6
drf-spectacular==0.14.0
djangorestframework-simplejwt==4.6.0
user-agents==2.2.0